Rename schema command to jsonschema
